fix: generic Into<i32> for lock-kind, not a visible cast (#66)
i32::from(...) at call sites still tripped clippy::useless_conversion
on Linux (where the constant is already i32) the same way `as i32`
tripped unnecessary_cast — any call-site conversion is a no-op on
whichever platform doesn't need it. fcntl_lock's kind parameter is
now `impl Into<i32>`, so the widening happens once, generically,
inside the function body; clippy can't flag a no-op it can't see
until monomorphization. Call sites pass the bare libc constant.

+++ src/vfs/shm.rs
@@ -123,12 +123,7 @@ impl Drop for WalReadLock {
     fn drop(&mut self) {
         // Best-effort: `drop` can't propagate a failure, and there is
         // nothing more this crate can do about one anyway.
-        let _ = fcntl_lock(
-            &self.file,
-            i32::from(libc::F_UNLCK),
-            wal_read_lock_byte(self.slot),
-            1,
-        );
+        let _ = fcntl_lock(&self.file, libc::F_UNLCK, wal_read_lock_byte(self.slot), 1);
     }
 }
 
@@ -166,10 +161,10 @@ pub(crate) fn claim_wal_read_lock(shm_path: &Path) -> io::Result<Option<WalReadL
         // Briefly exclusive, only long enough to publish this slot's mark
         // before downgrading to the SHARED lock held for the guard's
         // lifetime — matches SQLite's own claim sequence (spike 005 exp 4).
-        match fcntl_lock(&file, i32::from(libc::F_WRLCK), byte, 1) {
+        match fcntl_lock(&file, libc::F_WRLCK, byte, 1) {
             Ok(()) => {
                 set_read_mark(&file, slot, mx_frame(&file)?)?;
-                fcntl_lock(&file, i32::from(libc::F_RDLCK), byte, 1)?;
+                fcntl_lock(&file, libc::F_RDLCK, byte, 1)?;
                 return Ok(Some(WalReadLock { file, slot }));
             }
             Err(e) if matches!(e.raw_os_error(), Some(libc::EAGAIN) | Some(libc::EACCES)) => {
